refactor(workspace): extract useWorkspaceNavigation hook (1774→1646 lines)
Moves routing derivation, nav config building, secondary filter state, and
handleSecondaryFilterSelect into a dedicated hook.
